requirements.txt
# Requirements for Waste Detection Pipeline
# Core dependencies for training, inference, and data augmentation

# PyTorch ecosystem
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0

# Data processing and augmentation
albumentations>=1.3.0
numpy>=1.24.0

# COCO dataset tools
pycocotools>=2.0.6

# Visualization
matplotlib>=3.7.0

# ONNX support
onnx>=1.14.0
onnxruntime>=1.15.0

# OpenVINO support
openvino>=2025.0.0


